+++ src/media/setStorySeen.js
@@ -10,7 +10,7 @@
  */
 'use strict';
 
-var utils = require('../utils.js');
+var utils = require('../../utils');
 var log = require('npmlog');
 
 module.exports = function(defaultFuncs, api, ctx) {