+++ src/messaging/setMessageReactionMqtt.js
@@ -1,7 +1,7 @@
 
 'use strict';
 
-const { generateOfflineThreadingID } = require('../utils');
+const { generateOfflineThreadingID } = require('../../utils');
 
 function isCallable(func) {
   try {